update paraNumAnalysis, make paraTrajSegAnalysis.m
updated paraNumAnalysis.m to plot mean speed, MSD, and TSD. Also calculated Pearson correlation, plot it, and exported it as an image. Also applied machine learning (clustering) to see how many clusters are optimal for each dataset.
After these analyses, it has been concluded that NINV/INV is not a useful classification, nor the calculated data assuming that NINV/INV are the only classification are significant. Thus created paraTrajSegAnalysis.m to segment the trajectories for calculation and classification based on the smaller instances of locomotion.